Moved SaveSampleFromREDCap from sample_resource.py to a different package for easier management
