+++ api/__init__.py
@@ -66,7 +66,7 @@ def register_resources(app):
     from api.resources.index_resource import IndexResource
     from api.resources.auth_resource import AuthResource, LogOutResource
     from api.resources.theme_resource import ThemeResource
-    from api.resources.sample_resource import SampleResource, SaveSampleFromREDCap
+    from api.resources.sample_resource import SampleResource
     from api.resources.user_resource import UserResource
     from api.resources.publication_resource import PublicationResource
     from api.resources.box_resource import BoxResource
@@ -81,6 +81,7 @@ def register_resources(app):
     from api.resources.housedata_resource import HouseDataResource
     from api.resources.email_confirmation.email_confirmation import EmailConfirmationResource
     from api.resources.project_resource import ProjectResource
+    from .resources.redcap_requests.fetch_sample_resource import SaveSampleFromREDCap
 
     api = Api(app)
     api.add_resource(IndexResource, '/', '/index', '/welcome')
